requirements.txt
numpy==1.26.4
pillow==10.3.0
scikit-image==0.22.0
scipy==1.11.4
torch==2.3.1
torchvision==0.18.1

# SegFormer 
transformers==4.44.2
huggingface-hub>=0.23.0,<0.25
safetensors>=0.4.3
timm==0.9.12
requests>=2.31.0
packaging>=23.2
pyyaml>=6.0.1
